add staff done

+++ src/main/java/com/alpha5/autoaid/repository/UserRepository.java
@@ -3,7 +3,7 @@ package com.alpha5.autoaid.repository;
 import com.alpha5.autoaid.model.UserData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserData, Integer> {
+public interface UserRepository extends JpaRepository<UserData, Long> {
     UserData findByEmail(String email);
 
     UserData findByContactNo(String contactNo);
